rita/requirements.txt
flask==3.0.0
flask-login==0.6.3

google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
authlib==1.3.0

python-dotenv==1.0.0
pyyaml==6.0.1

gunicorn==21.2.0

requests==2.31.0
